requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
loguru==0.6.0
multidict==6.6.3
mysql-connector-python==9.3.0
propcache==0.3.2
pyami-asterisk==1.5.2
typing_extensions==4.14.1
yarl==1.20.1
